+++ web/WEB-INF/classes/src/com/quatro/service/security/UsersManager.java
@@ -67,7 +67,7 @@ public class UsersManager {
 		return secroleDao.getRoles();
 	}
 	public Secrole getRole(String id) {
-		return secroleDao.getRole(Long.valueOf(id));
+		return secroleDao.getRole(Integer.valueOf(id));
 	}
 	public Secrole getRoleByRolename(String roleName) {
 		return secroleDao.getRoleByName(roleName);
