+++ web/WEB-INF/classes/src/org/oscarehr/PMmodule/dao/ProgramClientStatusDAO.java
@@ -67,7 +67,7 @@ public class ProgramClientStatusDAO extends HibernateDaoSupport {
         }
 
         Query query = getSession().createQuery("select pt.id from ProgramClientStatus pt where pt.programId = ? and pt.name = ?");
-        query.setLong(0, programId.longValue());
+        query.setInteger(0, programId);
         query.setString(1, statusName);
 
         List teams = query.list();
