+++ web/WEB-INF/classes/src/org/oscarehr/PMmodule/model/FormFollowUp.java
@@ -204,7 +204,7 @@ public class FormFollowUp implements Serializable {
 
     private int hashCode = Integer.MIN_VALUE;// primary key
 
-    private Long _id;// fields
+    private Integer _id;// fields
     private String _cboxAssistwithhealthcard;
     private String _commentsOnEducation;
     private String _typeOfIncome;
@@ -333,7 +333,7 @@ public class FormFollowUp implements Serializable {
     private String _yearsOfEducation;
     private String _radioYourcanadianstatus;
     private String _radioDrugusefrequency;
-    private Long _providerNo;
+    private Integer _providerNo;
     private String _doctor1Phone;
     private String _agency1Name;
     private String _cboxHostel;
@@ -351,7 +351,7 @@ public class FormFollowUp implements Serializable {
     private String _needAssistInLegal;
     private String _radioCaredforschizophrenia;
     private java.util.Date _formCreated;
-    private Long _demographicNo;
+    private Integer _demographicNo;
     private String _radioDoyoudrink;
     private String _radioDrinking;
     private String _radioHealthproblem;
@@ -383,7 +383,7 @@ public class FormFollowUp implements Serializable {
      /**
       * Constructor for primary key
       */
-     public FormFollowUp (Long _id) {
+     public FormFollowUp (Integer _id) {
          this.setId(_id);
          initialize();
      }
@@ -392,8 +392,8 @@ public class FormFollowUp implements Serializable {
       * Constructor for required fields
       */
      public FormFollowUp (
-         Long _id,
-         Long _demographicNo) {
+    		 Integer _id,
+    		 Integer _demographicNo) {
 
          this.setId(_id);
          this.setDemographicNo(_demographicNo);
@@ -409,7 +409,7 @@ public class FormFollowUp implements Serializable {
 *  generator-class="native"
 *  column="ID"
 */
-    public Long getId () {
+    public Integer getId () {
         return _id;
     }
 
@@ -417,7 +417,7 @@ public class FormFollowUp implements Serializable {
 	 * Set the unique identifier of this class
      * @param _id the new ID
      */
-    public void setId (Long _id) {
+    public void setId (Integer _id) {
         this._id = _id;
         this.hashCode = Integer.MIN_VALUE;
     }
@@ -2345,7 +2345,7 @@ public class FormFollowUp implements Serializable {
     /**
 	 * Return the value associated with the column: provider_no
      */
-    public Long getProviderNo () {
+    public Integer getProviderNo () {
         return _providerNo;
     }
 
@@ -2353,7 +2353,7 @@ public class FormFollowUp implements Serializable {
 	 * Set the value related to the column: provider_no
      * @param _providerNo the provider_no value
      */
-    public void setProviderNo (Long _providerNo) {
+    public void setProviderNo (Integer _providerNo) {
         this._providerNo = _providerNo;
     }
 
@@ -2615,7 +2615,7 @@ public class FormFollowUp implements Serializable {
     /**
 	 * Return the value associated with the column: demographic_no
      */
-    public Long getDemographicNo () {
+    public Integer getDemographicNo () {
         return _demographicNo;
     }
 
@@ -2623,7 +2623,7 @@ public class FormFollowUp implements Serializable {
 	 * Set the value related to the column: demographic_no
      * @param _demographicNo the demographic_no value
      */
-    public void setDemographicNo (Long _demographicNo) {
+    public void setDemographicNo (Integer _demographicNo) {
         this._demographicNo = _demographicNo;
     }
