+++ web/WEB-INF/classes/src/org/oscarehr/PMmodule/model/FunctionalUserType.java
@@ -32,7 +32,7 @@ public class FunctionalUserType  implements Serializable {
     public static String PROP_NAME = "Name";
     public static String PROP_ID = "Id";
     private int hashCode = Integer.MIN_VALUE;// primary key
-    private Long _id;// fields
+    private Integer _id;// fields
     private String _name;
 
     // constructors
@@ -43,7 +43,7 @@ public class FunctionalUserType  implements Serializable {
 	/**
 	 * Constructor for primary key
 	 */
-	public FunctionalUserType (Long _id) {
+	public FunctionalUserType (Integer _id) {
 		this.setId(_id);
 		initialize();
 	}
@@ -58,7 +58,7 @@ public class FunctionalUserType  implements Serializable {
 *  generator-class="native"
 *  column="id"
 */
-    public Long getId () {
+    public Integer getId () {
         return _id;
     }
 
@@ -66,7 +66,7 @@ public class FunctionalUserType  implements Serializable {
 	 * Set the unique identifier of this class
      * @param _id the new ID
      */
-    public void setId (Long _id) {
+    public void setId (Integer _id) {
         this._id = _id;
         this.hashCode = Integer.MIN_VALUE;
     }
